+++ Tests/GUI/Model/GUIMod.cs
@@ -16,6 +16,7 @@ using CKAN.GUI;
 using CKAN.Versioning;
 
 using Tests.Core.Configuration;
+using Tests.Core.Relationships;
 using Tests.Data;
 
 namespace Tests.GUI
@@ -64,23 +65,54 @@ namespace Tests.GUI
                 using (var repo = new TemporaryRepository(old_version.ToJson(),
                                                           new_version.ToJson()))
                 using (var repoData = new TemporaryRepositoryData(user, repo.repo))
-            using (var cacheDir = new TemporaryDirectory())
-            using (var cache    = new NetModuleCache(cacheDir))
-            {
+                using (var cacheDir = new TemporaryDirectory())
+                using (var cache    = new NetModuleCache(cacheDir))
+                {
                     var registry = new Registry(repoData.Manager, repo.repo);
 
                     registry.RegisterModule(old_version, new List<string>(), tidy.KSP, false);
-                    var upgradeableGroups = registry.CheckUpgradeable(tidy.KSP,
-                                                                      new HashSet<string>());
+                    var upgradeable = registry.UpgradeableModules(tidy.KSP, new HashSet<string>()).ToArray();
 
                     var mod = new GUIMod(old_version, repoData.Manager, registry,
                                          tidy.KSP.StabilityToleranceConfig, tidy.KSP, cache,
                                          null, false, false)
                     {
-                        HasUpdate = upgradeableGroups[true].Any(m => m.identifier == old_version.identifier),
+                        HasUpdate = upgradeable.Any(m => m.identifier == old_version.identifier),
                     };
                     Assert.True(mod.HasUpdate);
                 }
+            }
+        }
+
+        [Test]
+        public void DownloadSizeInstallSize_Metapackage_NSlashA()
+        {
+            // Arrange
+            var metapackJson = RelationshipResolverTests.MergeWithDefaults(
+                                   @"{
+                                       ""identifier"": ""MyModPack"",
+                                       ""version"":    ""1.0"",
+                                       ""kind"":       ""metapackage""
+                                   }");
+            var user = new NullUser();
+            using (var gameInstWrapper = new DisposableKSP())
+            using (var repo = new TemporaryRepository(metapackJson))
+            using (var repoData = new TemporaryRepositoryData(user, repo.repo))
+            using (var cacheDir = new TemporaryDirectory())
+            using (var cache    = new NetModuleCache(cacheDir))
+            {
+                var metapack = CkanModule.FromJson(metapackJson);
+
+                // Act
+                var gmod = new GUIMod(metapack, new RepositoryDataManager(),
+                                                new CKAN.Registry(repoData.Manager, repo.repo),
+                                                gameInstWrapper.KSP.StabilityToleranceConfig,
+                                                gameInstWrapper.KSP, cache, false, false, false);
+
+                // Assert
+                Assert.AreEqual("N/A", gmod.DownloadSize);
+                Assert.AreEqual("N/A", gmod.InstallSize);
+                Assert.AreEqual("1.0", gmod.Version);
             }
         }